requirements.txt
# This file may be used to create an environment
python-telegram-bot=11.1.0
yaml=0.1.7
